Selector engine for the extension
The selector options operate very like the `rows()`, `row()`, etc
methods in DataTables whereby you can use a number of options to select
one or more buttons and then perform actions on them. The API is still
very simple here, but it will be fleshed out.

The DataTables API acts as the primary end developer interface to the
buttons library, but Buttons itself has a number of publich methods that
the API methods use - the Button API methods will not be publically
documented, prefering instead the interface that is added to the
DataTables API.

The Buttons "internal" API is index based for the buttons - each action
method must address a single method. It is up to the DataTables API
"bridge" to abstract that complexity out to allow manipulate by class
name, etc and for multiple buttons at a time.