Bootstrap support for Buttons
This has primarily involved adding the ability to have different button
tags / classes in the dropdown, compared to the top level. This is
currently undocumented, but likely I will add documentation for it
shortly.

The fixed layout is supported in Bootstrap (and will be for the other
integration frameworks). The example is somewhat basic atm...