Adding a button container option for better Foundation support
Foundation assumes that rather than each item in the list being the
button itself, is just a container for the button (makes sense for a
ul/li/a list). So for full support with Foundation this additional
option means that the container element is the one that is inserted, but
operations still happen on the button. It is optional.

Documentation updated to be much more simple - rather than redundantly
reproducing the tag and className documentation I've just defined this
options in the container object which is common to the dom options.