Print button - constructs a copy of the table based on the selected data, displaying it in a new window for printing.
This method is signficiantly different from TableTools which kept the
table in the page hiding all the surrounding elements. While that
generally worked okay, it doesn't allow for careful selection of the
data to be printed - columns and rows.

The disadvantage of doing it this new way is that the original table
retained its styling, since of course it was the original table! This
way row and cell classes are lost. If that becomes a signficiant problem
it could be resolved by reading that information from DataTables and
adding it to the constructed table.