Updating Flash PDF export to use JSON to transport the data rather than a messy self defined text format. This is _much_ cleaner in the AS3.
The format of the PDF table unfortunately can't be made to match the
table from the HTML5 table though due to the limitations of AlivePDF and
it hardly seems worth the effort to change that.

This is something that should probably have been done in TableTools a
long time ago!