updated function to accept individual columns as input
I updated the function, and applied it to the footers also.  

The function is now called for each column which is being exported and it is passed 3 values.  The 'colnum' is the column number where the data is from, the 'title' is the existing display title and the default response, and the 'column' is the column object.  Whatever string/value is returned from the function after the function runs is then used to populate the column header/footer.

If someone creates their own parse function for the column headers they can use the columnHeaders config setting to replace the default function with their custom function to custom parse the header data.