Attach hiddenDiv to table container instead of body
Older versions of IE do not allow a focus to be set to a hidden element.  This creates an issue with the copy function in older IE browsers where if a datatable is embedded in a modal or other element which overlays the entire page, the textarea in the hiddenDiv cannot receive the focus, so it's contents cannot be copied.  
By changing where we attach the hidden div so we are attaching it to the table container, we are ensuring that we can use the copy feature in older browsers.