Take line breaks into account when calculating column widths
When line breaks ( html br tag) are present in a cell, for example column 1 - product item, column 2 - multiple partnumbers (each around 16 characters), it's easy to convert them to text line breaks. For example:
format: {
	body: function( data, row, column, node){
		if(data.match(/&lt;br\s*\/?&gt;/gmi)){
			wrapColumns.push(column);
			//FOR IE, other browsers data.replace(/&lt;br\s*\/?&gt;/gmi, '\n'); ???
			return data.replace(/&lt;br&gt;/gmi, '\r\n');
		} else {
			//Strip html tags from data content
			return data.replace(/&lt;\/?[^&gt;]+(&gt;|$)/gmi, '');
		}
At the moment the calculation of the column width doesn't take these line breaks into account and in my example case will often create columns with width 40 and displaying them all in one line without separation. With the addition of the new styles in version 1.2.2 , one could apply the wrap text style and with this pull request the column width would be calculated and displayed correctly (through the bestFit=1 attribute added on line 1161).
Added config as parameter to internal function _excelColWidth for automatic detection of OS platform. Reason: decide which regex to apply (line 609, 618 and 620).
Slightly increased the len value by multiplying with 1.2 for aesthetic reasons (line 621 and 623). Adds a little more space.
Set default value to 8 instead of 5 (line 637), as it compares better to the standard value used by excel.